revert using unsigned integer

+++ encoders/hashgrid.py
@@ -31,7 +31,7 @@ class HashGridEncoder(nn.Module):
         self.resolutions = self.resolutions.reshape((-1, 1))
 
         # offset to the l-th hash table at the l-level
-        self.table_offsets = self.table_size * jnp.arange(self.num_levels, dtype=jnp.uint32)
+        self.table_offsets = self.table_size * jnp.arange(self.num_levels)
 
         self.hash_table = self.param("hash_table",
                                      lambda key, shape: jax.random.uniform(key, shape,
@@ -47,8 +47,8 @@ class HashGridEncoder(nn.Module):
         Returns:
              encoded: (num of points,) hash-encoded coordinates
         """
-        primes = jnp.array([1, 2_654_435_761, 805_459_861], dtype=jnp.uint32)
-        encoded = jax.lax.reduce(ijk * primes, jnp.uint32(0), jnp.bitwise_xor, dimensions=(2,))
+        primes = jnp.array([1, 2_654_435_761, 805_459_861])
+        encoded = jax.lax.reduce(ijk * primes, 0, jnp.bitwise_xor, dimensions=(2,))
         encoded = encoded % self.table_size
         return encoded
 
@@ -97,7 +97,7 @@ class HashGridEncoder(nn.Module):
         # (num points, L, 3) scaled x, y, z coordinates for each level
         scaled = jax.vmap(lambda res: points * res, out_axes=1)(self.resolutions)
         # (num points, L, 3) i, j, k indices of the first vertex.
-        floors = jnp.floor(scaled).astype(jnp.uint32)
+        floors = jnp.floor(scaled).astype(jnp.int32)
         weights = scaled - floors
 
         vertices_of_cell = jnp.asarray([
@@ -109,11 +109,10 @@ class HashGridEncoder(nn.Module):
             [1, 0, 1],
             [0, 1, 1],
             [1, 1, 1],
-        ], dtype=jnp.uint32)
+        ])
 
         # (num points, L, 8, 3), i, j, k coordinates of the 8 vertices where the input
         # point falls to. There are L levels of grids.
-        # TODO: use einsum?
         all_vertices = floors[:, :, jnp.newaxis, :] + vertices_of_cell[jnp.newaxis, jnp.newaxis, :]
 
         encoded_vertices = jax.vmap(for_each_layer,
